+++ package.json
@@ -109,6 +109,7 @@
     "@testing-library/react": "^11.0.4",
     "@testing-library/user-event": "^12.1.7",
     "@types/jest": "^25.1.2",
+    "@types/jest-axe": "^3.5.0",
     "@types/reach__alert": "^0.1.1",
     "@types/react": "16.8.1",
     "@types/react-dom": "16.8.1",
@@ -144,6 +145,7 @@
     "gatsby": "2.6.0",
     "husky": "^1.3.1",
     "jest": "^26.5.0",
+    "jest-axe": "^4.0.0",
     "jest-dom": "^3.1.1",
     "jest-emotion": "^10.0.14",
     "jest-environment-jsdom-sixteen": "^1.0.3",
