+++ setup-test-env.js
@@ -1,1 +1,2 @@
 import "@testing-library/jest-dom/extend-expect"
+import "jest-axe/extend-expect"
